move arguments.js into contracts
